fix: apply PR comment using decode hooks
